requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
redis>=5.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
anthropic>=0.40.0
sse-starlette>=2.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
